Linux Driver: fix a synchronization issue for mmput
mmput_async is required to avoid deadlock, but it is not exported for direct use in some older kernels.

Use kallsyms_lookup_name to resolve mmput_async for kernels below 5.4
Note: The mainline kernel does not export it, but newer distro kernels like Ubuntu kernel 5.4 do.
Note: The mainline kernel 5.7 and up do not export kallsyms_lookup_name, so this method can't be used for those kernels.

Signed-off-by: Haitao Huang <4699115+haitaohuang@users.noreply.github.com>